* more re-organization * adding chain handlers for limit/offset
